Clean up

+++ rgb-led-matrix/include/CFG/CFG.h
@@ -8,19 +8,19 @@
 namespace rgb_matrix {
     class CFG {
         public:
-            CFG(int rows, int cols, Data_Format_ID data_format, GAMMA gamma);
+            CFG(uint16_t rows, uint16_t cols, Data_Format_ID data_format, GAMMA gamma);
 
             DOTCorrect& get_dot();
             GAMMA& get_gamma();
-            int get_rows();
-            int get_cols();
+            uint16_t get_rows();
+            uint16_t get_cols();
             Data_Format_ID get_data_format();
         
         protected:
             CFG();
 
-            int rows_;
-            int cols_;
+            uint16_t rows_;
+            uint16_t cols_;
             DOTCorrect dot_;
             GAMMA gamma_;
             Data_Format_ID data_format_;
