+++ rgb-led-matrix/include/Mapper/Pixel/Pixel.h
@@ -2,7 +2,6 @@
 #define PIXEL_H
 
 #include <Panel/Single_Panel.h>
-#include <CFG/types.h>
 
 namespace rgb_matrix {
     // Optional construct for handling pixel locations and RGB ordering
@@ -11,7 +10,7 @@ namespace rgb_matrix {
             Pixel(Single_Panel *panel);
             virtual ~Pixel();
 
-            void SetPixel(int x, int y, uint8_t red, uint8_t green, uint8_t blue);
+            void SetPixel(uint16_t x, uint16_t y, uint8_t red, uint8_t green, uint8_t blue);
             void set_brightness(uint8_t brightness);
             void map_wavelength(uint8_t color, Color index, uint16_t value);
             void show(Protocol *protocol, bool schedule = true);
@@ -19,8 +18,8 @@ namespace rgb_matrix {
         protected:
             Pixel();
 
-            virtual cord_t map_location(int x, int y) = 0;
-            virtual Color_Order map_color(int x, int y) = 0;
+            virtual cord_t map_location(uint16_t x, uint16_t y) = 0;
+            virtual Color_Order map_color(uint16_t x, uint16_t y) = 0;
 
             Single_Panel *panel_;
