+++ rgb-led-matrix/include/RGBMatrix.h
@@ -1,7 +1,6 @@
 #ifndef RGBMATRIX_H
 #define RGBMATRIX_H
 
-#include <Panel/Single_Panel.h>
 #include <Panel/MultiPanel.h>
 #include <CFG/CFG.h>
 
@@ -10,7 +9,7 @@ namespace rgb_matrix {
     class RGBMatrix {
         public:
             static Single_Panel *Create_Single_Panel(CFG *cfg);
-            static MultiPanel *Create_MultiPanel(int width, int height);
+            static MultiPanel *Create_MultiPanel(uint16_t width, uint16_t height);
     };
 }
 #endif
