+++ rgb-led-matrix/lib/src/port/linux_x64/test/main.cpp
@@ -29,7 +29,7 @@ int main(int argc, char **argv) {
     Frame_Manager *manager = new Frame_Manager();
 
     // Draw (Double Buffered)
-    uint8_t x = 0, y = 0, i = 0;
+    uint16_t x = 0, y = 0, i = 0;
     while (1) {
         if (f[i]->isFree()) {
             f[i]->SetPixel(x, y, 255, 255, 255);
